refactor: restructure most of the code in five submodules
This commit introduces five submodules to clean the code: dca (distance of
closest approach), gang (based on the initials of the two authors),
initialization, shadowcasting and utils. Some .py files still need to be
repositioned, potentially adding an additional submodule, and functions
could be thematically grouped together.
