+++ pNeuma_simulator/batch.py
@@ -3,16 +3,18 @@ import warnings
 from joblib import Parallel
 
 from pNeuma_simulator import params
-from pNeuma_simulator.collision_exception import CollisionException
+from pNeuma_simulator.gang.collision_exception import CollisionException
 from pNeuma_simulator.main import main
 
 
 def batch(seed: int, permutation: tuple):
     """
     Run a batch simulation with the given seed and permutation.
+
     Args:
         seed (int): The seed for random number generation.
         permutation (tuple): A tuple containing the number of cars and motorcycles.
+
     Returns:
         tuple: A tuple containing the simulation results for cars and motorcycles.
     """
