+++ pNeuma_simulator/gang/__init__.py
@@ -1,0 +1,7 @@
+from gang.collision_exception import CollisionException  # noqa F401
+from gang.collisions import collisions  # noqa F401
+from gang.navigate import navigate  # noqa F401
+from gang.neighborhood import neighborhood  # noqa F401
+from gang.particle import Particle  # noqa F401
+
+__version__ = "0.0.0"
